Email Slicer - Python
In this application we will request an email from the user and slice the received email into domain and username. Imagine making an application that many people can use with the Python programming language. Each user needs an email and password to access their account. Most of the time, we want to keep the e-mail entered by the user separately as domain and user name. In this case, we make use of the index feature of the python programming language.